+++ flask_ai/resources/framework/app/constants.py
@@ -1,8 +1,10 @@
 RESPONSE_FOR_INVALID_QUERY = "I am a Resmed chatbot, I can't help with that"
+SLEEP_ASSESSMENT_URL = "https://www.resmed.com.au/online-sleep-assessment"
+
 
 # Link for sleep assessment
-SLEEP_ASSESSMENT_HTML_RESPONSE = "For more information please visit <a href='https://www.resmed.com.au/online-sleep-assessment' target='_blank'>Sleep Assessment</a><br/>"
-SLEEP_ASSESSMENT_RAW_RESPONSE = "For more information please visit Sleep Assessment - https://www.resmed.com.au/online-sleep-assessment"
+SLEEP_ASSESSMENT_HTML_RESPONSE = f"For more information please visit <a href='{SLEEP_ASSESSMENT_URL}' target='_blank'>Sleep Assessment</a><br/>"
+SLEEP_ASSESSMENT_RAW_RESPONSE = f"For more information please visit Sleep Assessment - {SLEEP_ASSESSMENT_URL}"
 
 
 '''